Started moving all HTML code outside PHP code. Fixed an error in the GPS conversion. Added a footer function to take over the "this file has been last edited on..." at the end of all forms, to standardize it, seeing how it was on all forms anyway. Working on changing the layout of the AP fetch page. Soon it will be:
---------------------------------------
  Assosiated list 
     |_> Signal history for that list 
       |_> GPS points for that Signal
---------------------------------------
Would like to have it so the GPS history is hideable via Javascript or something.
So far the changes are, all the previous seperate functions (fetch gps, fetch assosiate list, and fetch signal) are all now one function (fetch ap).
Unfortunately this is going to require a change in the backend. Previous databases will not be compatible with this update.
Reason being, in the previous versions (0.14 & 0.15[all builds]) did not store the signal history row in the Users table.
Old way ex 1,1-0,2-0,6-1,10-... / 0,6 : 0 is the update or new flag 1 = Updated AP 0 = New AP, the 6 is the AP ID number in the Database
New way ex 1,6:1-0,2:2-0,6:3-1,10:1-... /
0,6:2 ; 0 is the update or new flag 1 = Updated AP / 0 = New AP, the 6 is the Unique Access Point ID (UAPID) in the Database, and the 2 is the Signal History row number for the access point.)
The users table holds all the list imports for each user.
Fixed a bug when a search has no results, the page would output a PHP error, now it says "There where no results, please try again".
Fixed an issue, where on install there would be an SQL error and fail to install.
Added link to Last User on Index page.
Cleaned up the tables on the new version page.
Dates are now standardized as YYYY-MM-DD, to coincide with Vistumblers save file.
Moved the "*THIS PAGE* was last modified on: YYYY-MMM-DD @ HH:MM:SS" to the bottom of the page in the bottom cell.

git-svn-id: https://svn.code.sf.net/p/vistumbler/code@225 ecfccc70-5fee-4fae-9362-317d472445a1